pulled CountingHash methods back to Hashtable; removed counting.*
