+++ khmer/_khmer.cc
@@ -49,7 +49,6 @@ Contact: khmer-project@idyll.org
 #include "kmer_hash.hh"
 #include "hashtable.hh"
 #include "hashbits.hh"
-#include "counting.hh"
 #include "assembler.hh"
 #include "read_aligner.hh"
 #include "labelhash.hh"
